PythonAILib/python_ai_lib/test/requirements.txt
pillow
python-docx
python-dotenv
python-pptx
tiktoken
beautifulsoup4
chromadb
openai==1.52.1
magika
openpyxl
markdown
pdfminer.six
chardet
autogen-agentchat~=0.2
wikipedia
flaml[automl]
selenium
webdriver-manager
duckduckgo-search